WIP : Rename db.js as mongo.js
