+++ package.json
@@ -20,6 +20,7 @@
     "mongodb": "3.2.7",
     "node-cache": "4.2.0",
     "node-schedule": "1.3.2",
+    "pg": "^8.7.3",
     "snyk": "^1.161.1"
   },
   "keywords": [],
